+++ tests/unit/Driver/DriverConfigurationTest.php
@@ -15,13 +15,16 @@ class DriverConfigurationTest extends TestCase
         $this->assertNull($dc->getHttpProxy());
         $this->assertNull($dc->getFirefoxProfile());
         $this->assertSame('http://localhost:4444', $dc->getServerUrl());
+        $this->assertFalse($dc->shouldRunHeadless());
 
         $dc->setHttpProxy('0.0.0.1', 1234);
         $dc->setServerUrl('http://localhost:3333');
         $dc->setFirefoxProfile(new FirefoxProfile());
+        $dc->headlessMode();
 
         $this->assertInstanceOf(HttpProxy::class, $dc->getHttpProxy());
         $this->assertSame('http://localhost:3333', $dc->getServerUrl());
         $this->assertInstanceOf(FirefoxProfile::class, $dc->getFirefoxProfile());
+        $this->assertTrue($dc->shouldRunHeadless());
     }
 }